+++ packages/swr-login/CHANGELOG.md
@@ -1,5 +1,12 @@
 # swr-login
 
+## 0.7.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @swr-login/react@0.13.0
+
 ## 0.7.0
 
 ### Minor Changes